Remove usage legacy HTTP Client
Improved the "LookupUtils" to not use the deprecated HTTP classes
but "HttpURLConnection" instead. Contrary to the Apache HTTP client
the HttURLConnection is still existing in Android 6 and seems to be
more efficient besides being compatible with SSL SNI.

Furthermore set the missing "formattedNumber" attribute when the
"ContactBuilder" is creating the "ContactInfo" object after lookups.

Change-Id: I5ff9d96863d3d88ea2a6613020cc8667d49885a9

Conflicts:
	Android.mk

Conflicts:
	Android.mk
	AndroidManifest.xml